oops, short was not big enough to store command identifiers
this fixes the link selection menus in the about dialog

+++ src/about.cpp
@@ -78,7 +78,7 @@ void About::onInit()
 #endif
 }
 
-void About::onCommand(const short id, short)
+void About::onCommand(const int id, int)
 {
   switch(id) {
   case IDC_WEBSITE:
