+++ src/about.hpp
@@ -40,7 +40,7 @@ public:
 
 protected:
   void onInit() override;
-  void onCommand(short, short) override;
+  void onCommand(int, int) override;
   void onContextMenu(HWND, int x, int y) override;
 
 private:
