+++ src/browser.hpp
@@ -43,7 +43,7 @@ public:
 
 protected:
   void onInit() override;
-  void onCommand(short, short) override;
+  void onCommand(int, int) override;
   void onContextMenu(HWND, int x, int y) override;
   void onTimer(int) override;
