+++ src/import.hpp
@@ -34,7 +34,7 @@ public:
 
 protected:
   void onInit() override;
-  void onCommand(short, short) override;
+  void onCommand(int, int) override;
   void onTimer(int) override;
 
 private:
